Add PALRACE dataset
The PALRACE dataset were selected and adapted from the large-scale RACE dataset. We selected 800 high-school level questions from the test set of RACE and each question was associated with a distinct passage. The questions fell into 6 types, i.e., Cause (N = 200), Fact (N = 200), Inference (N = 120), Theme (N = 100), Title (N = 100), and Purpose (N = 80). The Cause, Fact, and Inference questions were concerned with the location, extraction, and comprehension of specific information from a passage, and were referred to as local questions. Questions of Theme, Title, and Purpose tested the understanding of a passage as a whole, and were referred to as global questions.